refactor: split xmlrpc/mod.rs into call/fault/parsing
Promotes xmlrpc/mod.rs (759 lines, cog 64) to:
  - mod.rs       re-exports + Value enum
  - call.rs      request building / param serialization
  - fault.rs     Fault parsing
  - parsing.rs   parse_value, parse_member, etc.

Public paths unchanged; xmlrpc/client.rs unchanged.
